ordersystem/requirements.txt
Django==2.0
django-redis==4.11.0
djangorestframework==3.11.0
mysqlclient==1.4.6
Pillow==7.0.0
pytz==2019.3
redis==3.4.1
